Make user decide what data type to use

+++ src/main/java/LibDL/example/ModuleExample.java
@@ -36,15 +36,15 @@ public class ModuleExample {
     }
 
     public static void main(String[] args) {
-        Variable data = new Constant(Nd4j.create(new double[][]{
-                {1.0, 0.0},
-                {1.0, 1.0},
-                {0.0, 1.0},
-                {0.0, 0.0}}));
+        Variable data = new Constant(Nd4j.create(new float[][]{
+                {1, 0},
+                {1, 1},
+                {0, 1},
+                {0, 0}}));
 
-        Variable target = new Constant(Nd4j.create(new double[][]{
-                {1.0}, {0.0},
-                {1.0}, {0.0}}));
+        Variable target = new Constant(Nd4j.create(new float[][]{
+                {1}, {0},
+                {1}, {0}}));
 
         Model nn = new Model();
         System.out.println(nn);
